calc: Intercept the input method activation event when inputting in the formula column
Since the formula bar will be sent to OxOffice every time you press a key, the content is displayed in the cell,
When the input method is activated, the roots will also be displayed, causing the phone or tablet to treat the roots as data.
So when the input method is enabled, it can be solved without displaying any data.