loleaflet: put the selected picture into the cache
Every time a picture is selected and dragged, it will download the SVG format of the picture,
Because it’s the same every time, it wastes resources and creates a picture download storm.
Now change to assign a hash id to each picture, and cache the picture data,
This only needs to be downloaded once, and this function requires patch OxOffice to work.
Otherwise, when you drag the graph, you can only see the dotted frame.