WOPI: Added WOPI storage error handling process.
When receiving status code 499 or 501 from WOPI,
The content of the response message must be sent to the front-end for processing.
Notify the user of the content of the information.

If the message content is in json format, the following two items are required:

{
"status": "status code(option)",
"message": "display message you want"
}

or plain text content.